+++ config/nvim/lua/config/lazy.lua
@@ -27,18 +27,18 @@ vim.opt.rtp:prepend(lazypath)
 require('lazy').setup({
   spec = {
     -- { import = "plugins.core" },
-    { import = "lazy.whichkey" },
-    { import = "lazy.ui" },
-    { import = "lazy.treesitter"},
-    { import = "lazy.lsp" },
-    { import = "lazy.avante" },
-    { import = "lazy.flutter" },
-    { import = "lazy.git" },
-    { import = "lazy.tool" },
-    { import = "lazy.oil" },
-    { import = "lazy.overseer" },
-    { import = "lazy.editor" },
-    { import = "lazy.skk" },
-    { import = "lazy.snippet" },
+    { import = "plugins.whichkey" },
+    { import = "plugins.ui" },
+    { import = "plugins.treesitter"},
+    { import = "plugins.lsp" },
+    { import = "plugins.avante" },
+    { import = "plugins.flutter" },
+    { import = "plugins.git" },
+    { import = "plugins.tool" },
+    { import = "plugins.oil" },
+    { import = "plugins.overseer" },
+    { import = "plugins.editor" },
+    { import = "plugins.skk" },
+    { import = "plugins.snippet" },
   },
 })
