+++ config/nvim/lua/plugins/avante.lua
