--- config/nvim/lua/plugins/chatgpt.lua
@@ -1,1 +1,0 @@
-require("chatgpt").setup()
