--- config/nvim/lua/plugins/copilot.lua
@@ -1,11 +1,0 @@
-local opts = { noremap = true, silent = true }
-local term_opts = { silent = true }
-
-local keymap = vim.api.nvim_set_keymap
-
--- vim.g.copilot_no_tab_map = true
-vim.g.copilot_assume_mapped = true
--- vim.g.copilot_tab_fallback = ""
--- vim.keymap.set('i', '<C-j>', '<Plug>(copilot-next)')
--- vim.keymap.set('i', '<C-k>', '<Plug>(copilot-previous)')
--- vim.keymap.set('i', '<C-/>', '<Plug>(copilot-dismiss)')
