--- config/nvim/lua/plugins/core.lua
@@ -1,1 +1,0 @@
-return { }
