--- config/nvim/lua/plugins/skkeleton.lua
@@ -1,7 +1,0 @@
-
-
-
--- vim.keymap.set({ "i", "c" }, [[<C-j>]], [[<Plug>(skkeleton-enable)]], opts)
---
---
--- require'skkeleton_indicator'.setup{}
