--- config/nvim/lua/plugins/toggleterm.lua
@@ -1,1 +1,0 @@
--- require("toggleterm").setup{}
