+++ config/nvim/lua/plugins/treesitter.lua
@@ -1,27 +1,53 @@
--- require'nvim-treesitter.configs'.setup {
---   highlight = {
---     enable = true,
---     disable = {},
---   },
---   indent = {
---     enable = false,
---     disable = {},
---   },
---   ensure_installed = {
---     "tsx",
---     "toml",
---     "json",
---     "yaml",
---     "swift",
---     "html",
---     "scss"
---   },
--- }
---
--- local vim = vim
--- local opt = vim.opt
---
--- opt.foldmethod = "expr"
--- opt.foldexpr = "nvim_treesitter#foldexpr()"
---
--- opt.foldenable = false
+return {
+  {
+    "nvim-treesitter/nvim-treesitter-context",
+    dependencies = { "nvim-treesitter/nvim-treesitter" },
+    opts = {
+      enable = true, -- Enable this plugin (Can be enabled/disabled later via commands)
+      multiwindow = false, -- Enable multiwindow support.
+      max_lines = 3, -- How many lines the window should span. Values <= 0 mean no limit.
+      min_window_height = 0, -- Minimum editor window height to enable context. Values <= 0 mean no limit.
+      line_numbers = true,
+      multiline_threshold = 20, -- Maximum number of lines to show for a single context
+      trim_scope = 'outer', -- Which context lines to discard if `max_lines` is exceeded. Choices: 'inner', 'outer'
+      mode = 'cursor',  -- Line used to calculate context. Choices: 'cursor', 'topline'
+      -- Separator between context and content. Should be a single character string, like '-'.
+      -- When separator is set, the context will only show up when there are at least 2 lines above cursorline.
+      separator = nil,
+      zindex = 20, -- The Z-index of the context window
+      on_attach = nil, -- (fun(buf: integer): boolean) return false to disable attaching
+    }
+  },
+  {
+    "nvim-treesitter/nvim-treesitter",
+    init = function()
+      vim.opt.foldmethod = "expr"
+      vim.opt.foldexpr = "nvim_treesitter#foldexpr()"
+
+      vim.opt.foldenable = false
+    end,
+    opts = {
+      highlight = {
+        enable = true,
+        disable = {},
+      },
+      indent = {
+        enable = false,
+        disable = {},
+      },
+      ensure_installed = {
+        "tsx",
+        "toml",
+        "json",
+        "yaml",
+        "swift",
+        "html",
+        "scss"
+      },
+    },
+    -- config = function()
+    --   require("plugins/treesitter")
+    -- end,
+    build = ":TSUpdate",
+  },
+}
