--- config/nvim/lua/plugins/trouble.lua
@@ -1,5 +1,0 @@
-require("trouble").setup {
-  -- your configuration comes here
-  -- or leave it empty to use the default settings
-  -- refer to the configuration section below
-}
