--- config/nvim/lua/plugins/ufo.lua
@@ -1,6 +1,0 @@
-
--- require('ufo').setup({
---     provider_selector = function(bufnr, filetype, buftype)
---         return {'treesitter', 'indent'}
---     end
--- })
